refactor(scripts): decompose retro-runner into phases/
Extracts the four retro pipeline phases plus their private helpers from
the monolithic retro-runner.js (932 LOC) into a sibling phases/
directory under .agents/scripts/lib/orchestration/retro/:

  - gather-signals.js     (305 LOC) — collectDescendants, friction sum,
                                       Epic body parser, body marker constants
  - compose-body.js       (290 LOC) — renderRoutedSections + normalizeInterventionCount
  - post-and-mirror.js    (140 LOC) — composeAndPostRetro orchestrator
  - checks.js             ( 94 LOC) — collectRetroFindings + rendering

The parent retro-runner.js shrinks to 182 LOC: owns the runRetro CLI
entry, arg validation, and the retro.start / retro.end bus boundary,
delegating body work to composeAndPostRetro. Re-exports gatherRetroSignals,
composeRetroBody, appendChecksSection, isCleanManifest, and the test
__INTERNAL_LABEL_REFERENCES so all call sites (epic-deliver-close-tail
plus 5 test files) keep importing from retro-runner.js unchanged.

Pure code-move refactor; no behavior change. 6751 tests pass (same
as base — 2 skipped are pre-existing).

Refs #3001 #2988
