+++ .agents/scripts/lib/orchestration/retro-runner.js
@@ -1,5 +1,5 @@
 /**
- * lib/orchestration/retro-runner.js — In-process Retro module.
+ * lib/orchestration/retro-runner.js — In-process Retro module (sequencer).
  *
  * Story #1155 (Epic #1142, 5.40.0) — extracts the helper-driven
  * `epic-retro` invocation into a callable module so the
@@ -10,10 +10,19 @@
  * opens the PR — the operator's PR-merge is the final human gate, not
  * the retro itself.
  *
+ * Story #3001 — the runner body was decomposed into `./retro/phases/`
+ * (one file per functional unit) following the established pattern at
+ * `./git-cleanup/phases/`, `./story-close/phases/`, and
+ * `./post-merge/phases/`. This file is now a thin sequencer that owns
+ * the `runRetro` entry point and re-exports the phase-level helpers so
+ * existing import sites stay unchanged.
+ *
  * Public API:
  *   - `runRetro({ epicId, provider, logger })` → `{ posted, compact, scorecard, body }`.
  *   - `composeRetroBody(input)` (pure, exported for tests).
  *   - `gatherRetroSignals({ epicId, provider })` (exported for tests).
+ *   - `appendChecksSection(body, findings)` (pure, exported for tests).
+ *   - `isCleanManifest` re-export.
  *
  * Behaviour:
  *   - Reads child Stories' `story-perf-summary` comments to aggregate
@@ -32,583 +41,20 @@
  */
 
 import nodeFs from 'node:fs';
-import path from 'node:path';
 
 import { runChecks } from '../checks/index.js';
 import { assembleState } from '../checks/state.js';
-import { epicRetroMirrorPath } from '../config/temp-paths.js';
-import { CONTEXT_LABELS, TYPE_LABELS } from '../label-constants.js';
-import { forEachLine } from '../observability/signals-writer.js';
+import { CONTEXT_LABELS } from '../label-constants.js';
+import { composeRetroBody } from './retro/phases/compose-body.js';
+import { gatherRetroSignals } from './retro/phases/gather-signals.js';
+import { composeAndPostRetro } from './retro/phases/post-and-mirror.js';
 import { isCleanManifest } from './retro-heuristics.js';
-import { composeRoutedProposals } from './retro-proposals.js';
-import { parseFencedJsonComment } from './structured-comment-parser.js';
-import { findStructuredComment, upsertStructuredComment } from './ticketing.js';
+import { upsertStructuredComment } from './ticketing.js';
 
-/**
- * Default framework repo (the Mandrel mirror) used when the caller does
- * not supply an override. Consumer projects re-routing framework friction
- * back to mandrel rely on this constant.
- */
-const DEFAULT_FRAMEWORK_REPO = 'dsj1984/mandrel';
-
-const RECUT_BODY_MARKER = /<!--\s*recut-of:\s*#?\d+\s*-->/;
-
-/**
- * Pure: aggregate `frictionByCategory` payloads into a single integer.
- */
-function sumFriction(byCategory) {
-  if (!byCategory || typeof byCategory !== 'object') return 0;
-  let total = 0;
-  for (const v of Object.values(byCategory)) {
-    if (typeof v === 'number' && Number.isFinite(v) && v >= 0) total += v;
-  }
-  return total;
-}
-
-/**
- * Walk every descendant ticket of `epicId` once. Returns the flat list with
- * each ticket's labels + body + state — the consumer derives its own
- * counts from this snapshot. Pure with respect to the provider injection.
- *
- * `provider.getSubTickets` is part of the `ITicketingProvider` contract;
- * the call is **not** optional-chained so a missing method (or a typo'd
- * test stub) surfaces as a thrown error rather than a silent empty graph.
- */
-async function collectDescendants(provider, epicId) {
-  const visited = new Set([epicId]);
-  const out = [];
-  // Story #2853 — level-order BFS so every parent at the current depth
-  // fires its getSubTickets call concurrently. Previously this loop
-  // awaited one parent at a time, serializing one network round-trip per
-  // descendant tier (a 2-level Epic with 10 Stories was 11 sequential
-  // calls). Each `getSubTickets` already fans out child hydration
-  // internally with concurrency=8 (issues.js:SUBTICKET_HYDRATION_CONCURRENCY),
-  // so outer parallelism is consistent with the existing design.
-  let frontier = [epicId];
-  while (frontier.length > 0) {
-    const results = await Promise.all(
-      frontier.map((id) => provider.getSubTickets(id)),
-    );
-    const nextFrontier = [];
-    for (const subs of results) {
-      for (const sub of subs ?? []) {
-        const subId = Number(sub?.id ?? sub?.number);
-        if (!Number.isInteger(subId) || visited.has(subId)) continue;
-        visited.add(subId);
-        out.push(sub);
-        nextFrontier.push(subId);
-      }
-    }
-    frontier = nextFrontier;
-  }
-  return out;
-}
-
-// Detects #NNN issue references in an Epic body — any match means the Epic
-// likely has child Stories enumerated in the planning artifact, so a
-// zero-descendant walk is suspicious.
-const EPIC_BODY_REFERENCE = /#\d+/;
-
-/**
- * Defensive guard: when the descendant walker returned zero, probe the
- * Epic ticket itself. If the Epic exists and its body references child
- * issues (e.g., `#123` planning refs), the empty walk is almost certainly
- * a contract drift in the underlying provider — emit a loud warn so the
- * silent failure surfaces. Probe errors degrade silently (the guard is
- * best-effort observability, not a correctness gate).
- */
-async function warnIfEpicLooksPopulated({ epicId, provider, logger }) {
-  let epic;
-  try {
-    epic = await provider.getTicket?.(epicId);
-  } catch (err) {
-    logger?.warn?.(
-      `[retro-runner] guard: getTicket(#${epicId}) failed (continuing): ${err?.message ?? err}`,
-    );
-    return;
-  }
-  if (!epic) return;
-  const body = typeof epic.body === 'string' ? epic.body : '';
-  if (!EPIC_BODY_REFERENCE.test(body)) return;
-  logger?.warn?.(
-    `[retro-runner] WARNING: Epic #${epicId} body references child issues, ` +
-      'but the descendant walker returned zero — retro will under-report ' +
-      'friction. Possible provider contract drift in getSubTickets.',
-  );
-}
-
-/**
- * Read raw retro signals from the GitHub graph. Pure with respect to the
- * provider injection — exported so tests can drive the predicate end-to-end
- * with a stub provider.
- *
- * When `descendants.length === 0`, the function probes `provider.getTicket`
- * to distinguish "Epic legitimately empty" from "the descendant walker
- * silently returned nothing under a populated Epic" (the failure mode the
- * pre-Story #2289 `getSubIssues` contract drift produced). A populated
- * Epic with zero walked descendants emits a loud warn via the supplied
- * `logger` so the silent failure becomes visible. Probe failure degrades
- * gracefully — the function never throws on the guard alone.
- *
- * @returns {Promise<{
- *   stories: Array<{ id: number, body?: string, labels?: string[] }>,
- *   tasks:   Array<{ id: number, labels?: string[] }>,
- *   counts:  { friction: number, parked: number, recuts: number, hotfixes: number, hitl: number },
- *   storyPerfSummaries: object[],
- *   epicPerfReport: object|null,
- *   parkedFollowOns: { recuts: object[], parked: object[], present: boolean },
- * }>}
- */
-export async function gatherRetroSignals({
-  epicId,
-  provider,
-  logger,
-  frameworkRepo,
-  consumerRepo,
-  forEachLineFn = forEachLine,
-  composeRoutedProposalsFn = composeRoutedProposals,
-}) {
-  const descendants = await collectDescendants(provider, epicId);
-  if (descendants.length === 0) {
-    await warnIfEpicLooksPopulated({ epicId, provider, logger });
-  }
-  const stories = descendants.filter((t) =>
-    (t.labels ?? []).includes(TYPE_LABELS.STORY),
-  );
-  const tasks = descendants.filter((t) =>
-    (t.labels ?? []).includes(TYPE_LABELS.TASK),
-  );
-
-  // Hotfix count: tasks that ever flipped to status::blocked. The label
-  // set on the closed Task is the cheapest signal (history would require
-  // event log access).
-  const hotfixes = tasks.filter((t) =>
-    (t.labels ?? []).includes('status::blocked'),
-  ).length;
-
-  // HITL count: distinct descendants that currently or historically carry
-  // `agent::blocked`. We can only see "currently" here without an event
-  // stream — counts undercount but never overcount.
-  const hitl = descendants.filter((t) =>
-    (t.labels ?? []).includes('agent::blocked'),
-  ).length;
-
-  // Aggregate per-Story `story-perf-summary` payloads for friction totals.
-  // Story #2853 — fan out the per-Story `findStructuredComment` lookups
-  // concurrently. Each Story's lookup resolves to a single paginated
-  // `getTicketComments` round-trip (the raw-comments cache at
-  // ticketing/reads.js:381-412 dedupes the per-type lookups within a
-  // ticket), so parallelization adds no extra wire calls — it just
-  // collapses N sequential round-trips into one wall-clock fan-out.
-  // `Promise.all` preserves index order, so the resulting
-  // `storyPerfSummaries` array stays deterministic relative to `stories`.
-  const perStoryComments = await Promise.all(
-    stories.map((story) =>
-      findStructuredComment(
-        provider,
-        story.id ?? story.number,
-        'story-perf-summary',
-      ),
-    ),
-  );
-  const storyPerfSummaries = [];
-  let frictionFromSummaries = 0;
-  for (const comment of perStoryComments) {
-    const parsed = parseFencedJsonComment(comment);
-    if (parsed) {
-      storyPerfSummaries.push(parsed);
-      frictionFromSummaries += sumFriction(parsed.frictionByCategory);
-    }
-  }
-
-  // Epic-level perf report (used by the full retro's "Top hotspots").
-  const epicPerfComment = await findStructuredComment(
-    provider,
-    epicId,
-    'epic-perf-report',
-  );
-  const epicPerfReport = parseFencedJsonComment(epicPerfComment);
-
-  // Parked + recut counts: prefer the structured comment; fall back to body
-  // grep for the recut marker when the comment is absent.
-  const parkedComment = await findStructuredComment(
-    provider,
-    epicId,
-    'parked-follow-ons',
-  );
-  const parkedParsed = parseFencedJsonComment(parkedComment);
-  let parkedFollowOns;
-  if (parkedParsed) {
-    parkedFollowOns = {
-      present: true,
-      recuts: Array.isArray(parkedParsed.recuts) ? parkedParsed.recuts : [],
-      parked: Array.isArray(parkedParsed.parked) ? parkedParsed.parked : [],
-    };
-  } else {
-    const recutsByBody = stories.filter(
-      (s) => typeof s.body === 'string' && RECUT_BODY_MARKER.test(s.body),
-    );
-    parkedFollowOns = {
-      present: false,
-      recuts: recutsByBody.map((s) => ({ storyId: s.id ?? s.number })),
-      parked: [],
-    };
-  }
-
-  const counts = {
-    friction: frictionFromSummaries,
-    parked: parkedFollowOns.parked.length,
-    recuts: parkedFollowOns.recuts.length,
-    hotfixes,
-    hitl,
-  };
-
-  // Story #2558 — read per-Story `signals.ndjson` streams (already
-  // source-tagged by Story #2553's writer) and compose the four routed
-  // proposal sections (framework / consumer / memory / discarded). Read
-  // failures degrade silently — observability MUST NOT take down the
-  // retro path. Empty streams yield empty arrays so the composer
-  // remains backward-compatible.
-  const routedSignals = [];
-  const memorablePatterns = [];
-  for (const story of stories) {
-    const sid = Number(story.id ?? story.number);
-    if (!Number.isInteger(sid) || sid <= 0) continue;
-    try {
-      await forEachLineFn(epicId, sid, (parsed) => {
-        if (parsed === null || typeof parsed !== 'object') return;
-        const record = /** @type {Record<string, unknown>} */ (parsed);
-        const category =
-          typeof record.category === 'string' ? record.category : null;
-        const source = record.source === 'framework' ? 'framework' : 'consumer';
-        if (category) {
-          routedSignals.push({ category, source });
-        }
-        if (record.memorable === true && typeof record.insight === 'string') {
-          memorablePatterns.push({
-            category: category ?? 'general',
-            insight: record.insight,
-          });
-        }
-      });
-    } catch (err) {
-      logger?.warn?.(
-        `[retro-runner] forEachLine failed for story #${sid} (continuing): ${
-          err?.message ?? err
-        }`,
-      );
-    }
-  }
-
-  // Resolve repos. Caller overrides win; otherwise default the consumer
-  // repo to the project's own `github.owner/repo` (best-effort: the
-  // provider may expose it, but we don't depend on it — empty string
-  // disables that pane in the routed proposals).
-  const resolvedFrameworkRepo =
-    typeof frameworkRepo === 'string' && frameworkRepo.length > 0
-      ? frameworkRepo
-      : DEFAULT_FRAMEWORK_REPO;
-  const resolvedConsumerRepo =
-    typeof consumerRepo === 'string' && consumerRepo.length > 0
-      ? consumerRepo
-      : resolvedFrameworkRepo; // when caller omits, fall back to the framework repo so the command renders without an empty `--repo` flag.
-
-  const routedProposals = composeRoutedProposalsFn({
-    epicId,
-    frameworkRepo: resolvedFrameworkRepo,
-    consumerRepo: resolvedConsumerRepo,
-    signals: routedSignals,
-    unresolvedBlockedEvents: [],
-    memorablePatterns,
-  });
-
-  return {
-    stories,
-    tasks,
-    counts,
-    storyPerfSummaries,
-    epicPerfReport,
-    parkedFollowOns,
-    routedProposals,
-  };
-}
-
-/**
- * Pure: clamp a candidate count to a non-negative integer. Used to
- * normalize the `manualInterventions` count plumbed in from the
- * epic-run-state-store snapshot before it lands in the scorecard.
- * Non-finite, negative, or non-numeric values collapse to 0.
- */
-function normalizeInterventionCount(value) {
-  if (typeof value !== 'number' || !Number.isFinite(value) || value < 0) {
-    return 0;
-  }
-  return Math.trunc(value);
-}
-
-/**
- * Pure: compose the retro markdown body. Exported for tests so they can
- * verify the body shape without round-tripping through a stub provider.
- *
- * Story #2289 adds `counts.interventions` — sourced from the
- * `manualInterventions` array on the `epic-run-state` snapshot (the
- * same list that disqualifies an Epic from auto-merge). Non-zero
- * interventions route to the full retro shape via `isCleanManifest`.
- *
- * @param {{
- *   epicId: number,
- *   epicTitle?: string,
- *   counts: { friction: number, parked: number, recuts: number, hotfixes: number, hitl: number, interventions?: number },
- *   storyPerfSummaries?: object[],
- *   epicPerfReport?: object|null,
- *   parkedFollowOns?: { recuts: object[], parked: object[] },
- *   tasksTotal?: number,
- *   tasksFirstTry?: number,
- *   timestamp?: string,
- *   forceFull?: boolean,
- * }} input
- * @returns {{ body: string, compact: boolean, scorecard: object }}
- */
-export function composeRetroBody(input) {
-  const {
-    epicId,
-    epicTitle = `Epic ${epicId}`,
-    counts,
-    epicPerfReport = null,
-    parkedFollowOns = { recuts: [], parked: [] },
-    routedProposals = null,
-    tasksTotal = 0,
-    tasksFirstTry = 0,
-    timestamp = new Date().toISOString(),
-    forceFull = false,
-  } = input;
-
-  const interventions = normalizeInterventionCount(counts?.interventions);
-  const compact = !forceFull && isCleanManifest({ ...counts, interventions });
-  const heading = `## 🪞 Sprint Retrospective — Epic #${epicId}: ${epicTitle}`;
-  const generatedLine = `_Generated ${timestamp}_`;
-  const scorecardRows = [
-    `| Total Tasks                  | ${tasksTotal} |`,
-    `| Tasks Completed First Try    | ${tasksFirstTry} |`,
-    `| Tasks Requiring Hotfix       | ${counts.hotfixes} |`,
-    `| agent::blocked Events Raised | ${counts.hitl} |`,
-    `| Manual Interventions         | ${interventions} |`,
-    `| Friction Events              | ${counts.friction} |`,
-  ];
-  const scorecard = {
-    totalTasks: tasksTotal,
-    tasksFirstTry,
-    hotfixes: counts.hotfixes,
-    hitl: counts.hitl,
-    friction: counts.friction,
-    parked: counts.parked,
-    recuts: counts.recuts,
-    interventions,
-  };
-  const completeMarker = `<!-- retro-complete: ${timestamp} -->`;
-
-  if (compact) {
-    const body = [
-      heading,
-      '',
-      generatedLine,
-      '',
-      '🟢 Clean sprint — zero friction, zero parked follow-ons, zero recuts, zero hotfixes, zero agent::blocked events, zero manual interventions.',
-      '',
-      '### Sprint Scorecard',
-      '',
-      '| Metric                       | Value |',
-      '| ---------------------------- | ----- |',
-      ...scorecardRows,
-      '',
-      '### Session Observations',
-      '',
-      '_Nothing notable beyond the scorecard._',
-      '',
-      '### Action Items for Next Epic',
-      '',
-      '_None._',
-      '',
-      completeMarker,
-    ].join('\n');
-    return { body, compact: true, scorecard };
-  }
-
-  // Full path — six sections.
-  const hotspotLines =
-    epicPerfReport &&
-    Array.isArray(epicPerfReport.topHotspots) &&
-    epicPerfReport.topHotspots.length > 0
-      ? epicPerfReport.topHotspots.map(
-          (h) =>
-            `- \`${h.phase}\` — ${h.occurrences} occurrence(s), avg ratio ${
-              typeof h.avgRatio === 'number' ? h.avgRatio.toFixed(2) : 'n/a'
-            }`,
-        )
-      : ['_No epic-perf-report available._'];
-
-  const parkedLines =
-    parkedFollowOns.parked.length > 0
-      ? parkedFollowOns.parked.map(
-          (p) =>
-            `- Adopt or close parked follow-on #${p.storyId ?? p.id ?? '?'}`,
-        )
-      : [];
-  const recutLines =
-    parkedFollowOns.recuts.length > 0
-      ? parkedFollowOns.recuts.map(
-          (r) => `- Recut #${r.storyId ?? r.id ?? '?'} attributed to manifest`,
-        )
-      : [];
-  const legacyActionItems = [...parkedLines, ...recutLines];
-  const legacyActionItemsBody =
-    legacyActionItems.length > 0 ? legacyActionItems.join('\n') : '_None._';
-
-  // Story #2558 — routed-proposals mode. When routedProposals is supplied
-  // AND any of the four buckets is non-empty, render the four explicit
-  // sections in deterministic order ABOVE the retro-complete marker:
-  //   1. Proposed issues — consumer repo
-  //   2. Proposed issues — framework repo
-  //   3. Proposed memory updates
-  //   4. One-off / discarded
-  // Otherwise the legacy "Action Items for Next Epic" section renders.
-  const routedSectionsBlock = renderRoutedSections(routedProposals);
-
-  const actionSection =
-    routedSectionsBlock === null
-      ? ['### Action Items for Next Epic', '', legacyActionItemsBody]
-      : routedSectionsBlock;
-
-  const body = [
-    heading,
-    '',
-    generatedLine,
-    '',
-    '### Sprint Scorecard',
-    '',
-    '| Metric                       | Value |',
-    '| ---------------------------- | ----- |',
-    ...scorecardRows,
-    '',
-    '### What Went Well',
-    '',
-    '_(populate from execution telemetry — extracted retro module emits a placeholder; deeper analysis is the operator follow-up.)_',
-    '',
-    '### What Could Be Improved',
-    '',
-    '#### Top hotspots',
-    '',
-    ...hotspotLines,
-    '',
-    '### Architectural Debt',
-    '',
-    '_(no automated detection in v5.40.0 — operator review required.)_',
-    '',
-    '### Protocol Optimization Recommendations (Self-Healing)',
-    '',
-    '_(operator follow-up.)_',
-    '',
-    ...actionSection,
-    '',
-    completeMarker,
-  ].join('\n');
-  return { body, compact: false, scorecard };
-}
-
-/**
- * Pure: render the four routed-proposal sections in deterministic order.
- * Returns `null` when `routedProposals` is absent or fully empty — the
- * caller falls back to the legacy "Action Items for Next Epic" section so
- * back-compat callers see no shape change.
- *
- * @param {{ framework: object[], consumer: object[], memory: object[], discarded: object[] } | null} routedProposals
- * @returns {string[] | null}
- */
-function renderRoutedSections(routedProposals) {
-  if (
-    !routedProposals ||
-    typeof routedProposals !== 'object' ||
-    Array.isArray(routedProposals)
-  ) {
-    return null;
-  }
-  const framework = Array.isArray(routedProposals.framework)
-    ? routedProposals.framework
-    : [];
-  const consumer = Array.isArray(routedProposals.consumer)
-    ? routedProposals.consumer
-    : [];
-  const memory = Array.isArray(routedProposals.memory)
-    ? routedProposals.memory
-    : [];
-  const discarded = Array.isArray(routedProposals.discarded)
-    ? routedProposals.discarded
-    : [];
-  if (
-    framework.length === 0 &&
-    consumer.length === 0 &&
-    memory.length === 0 &&
-    discarded.length === 0
-  ) {
-    return null;
-  }
-
-  const out = [];
-  out.push('### Proposed issues — consumer repo');
-  out.push('');
-  if (consumer.length === 0) {
-    out.push('_None._');
-  } else {
-    for (const item of consumer) {
-      out.push(`- **${item.title ?? item.category}**`);
-      out.push('');
-      out.push('```sh');
-      out.push(String(item.command ?? ''));
-      out.push('```');
-      out.push('');
-    }
-  }
-  out.push('');
-  out.push('### Proposed issues — framework repo');
-  out.push('');
-  if (framework.length === 0) {
-    out.push('_None._');
-  } else {
-    for (const item of framework) {
-      out.push(`- **${item.title ?? item.category}**`);
-      out.push('');
-      out.push('```sh');
-      out.push(String(item.command ?? ''));
-      out.push('```');
-      out.push('');
-    }
-  }
-  out.push('');
-  out.push('### Proposed memory updates');
-  out.push('');
-  if (memory.length === 0) {
-    out.push('_None._');
-  } else {
-    out.push('update your memory with the following insights:');
-    out.push('');
-    for (const m of memory) {
-      out.push(`- ${m.insight}`);
-    }
-  }
-  out.push('');
-  out.push('### One-off / discarded');
-  out.push('');
-  if (discarded.length === 0) {
-    out.push('_None._');
-  } else {
-    for (const d of discarded) {
-      out.push(
-        `- \`${d.category}\` (${d.occurrences ?? 1} occurrence, source: ${d.source ?? 'consumer'})`,
-      );
-    }
-  }
-  return out;
-}
+// Re-export phase-level helpers so existing import sites stay unchanged.
+export { appendChecksSection } from './retro/phases/checks.js';
+export { composeRetroBody } from './retro/phases/compose-body.js';
+export { gatherRetroSignals } from './retro/phases/gather-signals.js';
 
 /**
  * Public: compose and post the retro structured comment on the Epic.
@@ -723,202 +169,6 @@ export async function runRetro(opts = {}) {
     await bus.emit('retro.end', payload);
     throw err;
   }
-}
-
-/**
- * Inner compose-and-post helper. Extracted so `runRetro` can wrap the
- * full body in a try/catch for the `retro.end` boundary emit without
- * cluttering the happy-path read.
- */
-async function composeAndPostRetro({
-  epicId,
-  provider,
-  logger,
-  forceFull,
-  timestamp,
-  bus,
-  now,
-  manualInterventions,
-  gatherFn,
-  composeFn,
-  upsertFn,
-  runChecksFn,
-  assembleStateFn,
-  cwd,
-  fsImpl,
-  startedAt,
-  onMirrorWritten,
-}) {
-  const signals = await gatherFn({ epicId, provider, logger });
-
-  // Best-effort fetch of the Epic title for the heading.
-  let epicTitle;
-  try {
-    const epic = await provider.getTicket?.(epicId);
-    epicTitle = epic?.title;
-  } catch (err) {
-    logger?.warn?.(
-      `[retro-runner] Failed to fetch Epic #${epicId} title (using fallback): ${err?.message ?? err}`,
-    );
-  }
-
-  // tasksTotal: every Task descendant. tasksFirstTry: tasks that did not
-  // require a hotfix. (Conservative undercount when the runtime never
-  // flipped status::blocked, but the metric is honest.)
-  const tasksTotal = signals.tasks.length;
-  const hotfixCount = signals.counts.hotfixes;
-  const tasksFirstTry = Math.max(0, tasksTotal - hotfixCount);
-
-  const interventions = normalizeInterventionCount(manualInterventions);
-  const { body, compact, scorecard } = composeFn({
-    epicId,
-    epicTitle,
-    counts: { ...signals.counts, interventions },
-    storyPerfSummaries: signals.storyPerfSummaries,
-    epicPerfReport: signals.epicPerfReport,
-    parkedFollowOns: signals.parkedFollowOns,
-    routedProposals: signals.routedProposals,
-    tasksTotal,
-    tasksFirstTry,
-    timestamp,
-    forceFull,
-  });
-
-  const findings = await collectRetroFindings({
-    runChecksFn,
-    assembleStateFn,
-    cwd,
-    logger,
-  });
-  const bodyWithChecks = appendChecksSection(body, findings);
-
-  logger?.info?.(
-    `[retro-runner] Posting ${compact ? 'compact' : 'full'} retro on Epic #${epicId}${findings.length > 0 ? ` (${findings.length} finding(s))` : ''}...`,
-  );
-  const result = await upsertFn(provider, epicId, 'retro', bodyWithChecks);
-
-  // Story #2089: also mirror the retro body to the per-Epic temp dir so
-  // operators can read it locally without re-fetching from GitHub. GitHub
-  // remains SSOT — a write failure logs a warn and does not fail the
-  // phase. The path is resolved relative to `cwd` when supplied so that
-  // worktree-scoped invocations land under the worktree's temp tree.
-  let mirrorAbsPath = null;
-  try {
-    const rel = epicRetroMirrorPath(epicId);
-    const absPath = path.isAbsolute(rel)
-      ? rel
-      : path.join(cwd ?? process.cwd(), rel);
-    fsImpl.mkdirSync(path.dirname(absPath), { recursive: true });
-    fsImpl.writeFileSync(absPath, bodyWithChecks, 'utf8');
-    mirrorAbsPath = absPath;
-    onMirrorWritten?.(absPath);
-    logger?.info?.(`[retro-runner] Mirrored retro to ${absPath}`);
-  } catch (err) {
-    logger?.warn?.(
-      `[retro-runner] Failed to write retro mirror (retro.md) for Epic #${epicId} (continuing — GitHub upsert succeeded): ${err?.message ?? err}`,
-    );
-  }
-
-  // Story #2252 — emit `retro.end` after the upsert + mirror settle so
-  // the lifecycle ledger captures the closing boundary with the
-  // posted/compact flags AND the resolved mirror path (when present).
-  const endedAt = typeof now === 'function' ? now() : Date.now();
-  const retroEndPayload = {
-    epicId,
-    posted: true,
-    compact: Boolean(compact),
-    durationMs: Math.max(0, Math.floor(endedAt - startedAt)),
-  };
-  if (mirrorAbsPath) retroEndPayload.retroPath = mirrorAbsPath;
-  await bus.emit('retro.end', retroEndPayload);
-
-  return {
-    posted: true,
-    compact,
-    scorecard,
-    body: bodyWithChecks,
-    findings,
-    commentId: result?.commentId,
-  };
-}
-
-/**
- * Story #1290: invoke the self-healing checks registry with scope:'retro'
- * (read-only by construction — `runChecks` throws on autoFix:true under
- * this scope). Failures degrade gracefully to an empty findings list so
- * the retro never blocks on a probe error.
- */
-async function collectRetroFindings({
-  runChecksFn,
-  assembleStateFn,
-  cwd,
-  logger,
-}) {
-  try {
-    const state = await assembleStateFn({ scope: 'retro', cwd });
-    const result = await runChecksFn({
-      scope: 'retro',
-      autoFix: false,
-      state,
-    });
-    return Array.isArray(result?.findings) ? result.findings : [];
-  } catch (err) {
-    logger?.warn?.(
-      `[retro-runner] runChecks(scope:'retro') failed (continuing with empty findings): ${err?.message ?? err}`,
-    );
-    return [];
-  }
-}
-
-/**
- * Pure: append a "Self-Healing Checks" section to the retro body, listing
- * each finding's id, severity, summary, and fixCommand. When `findings` is
- * empty, the body is returned unchanged — this preserves the compact
- * "🟢 Clean sprint" shape under a clean manifest.
- *
- * The section is inserted **before** the `<!-- retro-complete: ... -->`
- * terminating marker so the marker stays at the end of the body (the
- * deliver pipeline searches for it as the EOF sentinel).
- *
- * Output format mirrors `/diagnose`'s renderTable for fixCommand display:
- * the same literal shell command appears verbatim in a fenced code block
- * so operators can copy-paste it.
- *
- * @param {string} body
- * @param {Array<import('../checks/index.js').Finding>} findings
- * @returns {string}
- */
-export function appendChecksSection(body, findings) {
-  if (!Array.isArray(findings) || findings.length === 0) return body;
-  const section = renderFindingsSection(findings);
-  const markerRe = /(<!--\s*retro-complete:[^>]*-->\s*)$/;
-  if (markerRe.test(body)) {
-    return body.replace(markerRe, `${section}\n$1`);
-  }
-  return `${body}\n${section}`;
-}
-
-function escapeCell(value) {
-  return String(value ?? '').replace(/\|/g, '\\|');
-}
-
-function renderFindingRow(f) {
-  const id = escapeCell(f?.id);
-  const severity = escapeCell(f?.severity);
-  const summary = escapeCell(f?.summary);
-  const fixCommand = escapeCell(f?.fixCommand);
-  return `| ${id} | ${severity} | ${summary} | \`${fixCommand}\` |`;
-}
-
-function renderFindingsSection(findings) {
-  return [
-    '### Self-Healing Checks',
-    '',
-    '| ID | Severity | Summary | Fix Command |',
-    '| --- | --- | --- | --- |',
-    ...findings.map(renderFindingRow),
-    '',
-  ].join('\n');
 }
 
 // Re-export for downstream test convenience — keeps the module's public
